fixed bit helpers

+++ src/helpers.h
@@ -11,7 +11,7 @@
  * @param num - The number of bits to read.
  * @return The read bits as a byte.
  */
-uint8_t lwmqtt_read_bits(uint8_t byte, uint8_t pos, uint8_t num);
+uint8_t lwmqtt_read_bits(uint8_t byte, int pos, int num);
 
 /**
  * Write bits to a byte.
@@ -21,7 +21,7 @@ uint8_t lwmqtt_read_bits(uint8_t byte, uint8_t pos, uint8_t num);
  * @param pos - The position of the first bit.
  * @param num - The number of bits to write.
  */
-void lwmqtt_write_bits(uint8_t *byte, uint8_t value, uint8_t pos, uint8_t num);
+void lwmqtt_write_bits(uint8_t *byte, uint8_t value, int pos, int num);
 
 /**
  * Reads arbitrary data from the specified buffer. The pointer is incremented by bytes read.
